server/requirements.txt
aerich==0.3.2
aiohttp==3.6.3
aiohttp-cors==0.7.0
aiohttp-devtools==0.13.1
aiomysql==0.0.20
aiosqlite==0.16.0
async-timeout==3.0.1
attrs==20.2.0
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
devtools==0.6
gmqtt==0.6.7
idna==2.10
iso8601==0.1.13
multidict==4.7.6
pycparser==2.20
pydantic==1.7.2
PyGithub>=1.53
Pygments==2.7.1
PyMySQL==0.9.2
PyPika==0.43.0
python-dotenv==0.15.0
requests>=2.22.0
six==1.15.0
tortoise-orm==0.16.17
typing-extensions==3.7.4.3
watchgod==0.6
websocket-client==0.57.0
yarl==1.5.1
